[✖️remove] : 불필요한 모듈 및 파일 삭제
