+++ src/admin/dto/index.ts
@@ -1,3 +1,3 @@
 export * from './suspension-user.dto';
-export * from '../interfaces/user-info-interface';
 export * from './deletion-user.dto';
+export * from './approval-user.dto';
