+++ src/admin/entities/index.ts
@@ -1,4 +1,4 @@
 export * from './deleted-users.entity';
-export * from './report-comments.entity';
-export * from './report-posts.entity';
+export * from '../../reports/entities/report-comments.entity';
+export * from '../../reports/entities/report-posts.entity';
 export * from './suspended-users.entity';
