+++ src/admin/interfaces/index.ts
@@ -1,5 +1,5 @@
-export * from './approval-user-list.interface'
-export * from './user-info-interface'
-export * from './user-list.interface'
-export * from './post-list.interface'
-export * from './comment-reply.interface'
+export * from './approval-user-list.interface';
+export * from './user-info-interface';
+export * from './user-list.interface';
+export * from './post-list.interface';
+export * from './comment-reply.interface';
