+++ src/admin/interfaces/post-list.interface.ts
@@ -1,7 +1,7 @@
 export interface IPostList {
-    postId: number; // 게시물 ID
-    boardType: string; // 카테고리
-    title: string; // 제목
-    author: string; // 작성자(닉네임)
-    createdAt: Date; // 작성일
+  postId: number; // 게시물 ID
+  boardType: string; // 카테고리
+  title: string; // 제목
+  author: string; // 작성자(닉네임)
+  createdAt: Date; // 작성일
 }
