+++ src/comments/comments.module.ts
@@ -6,7 +6,7 @@ import { CommentsEntity } from './entities/comments.entity';
 import { RepliesEntity } from '../replies/entities/replies.entity';
 import { PostsModule } from '../posts/posts.module';
 import { PostsEntity } from '../posts/entities/base-posts.entity';
-import { ReportCommentsEntity } from '../admin/entities/report-comments.entity';
+import { ReportCommentsEntity } from '../reports/entities/report-comments.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([CommentsEntity, RepliesEntity, PostsEntity, ReportCommentsEntity]), PostsModule],
