+++ src/common/interfaces/index.ts
@@ -1,2 +1,3 @@
 export * from './base-exception-interface';
 export * from './redis-options-interface';
+export * from './paginated-response-interface'
