--- src/hospitals/hospitals.module.ts
@@ -1,4 +1,0 @@
-import { Module } from '@nestjs/common';
-
-@Module({})
-export class HospitalsModule {}
