--- src/ocr/ocr.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { OcrController } from './ocr.controller';
-
-describe('OcrController', () => {
-  let controller: OcrController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [OcrController],
-    }).compile();
-
-    controller = module.get<OcrController>(OcrController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
