--- src/ocr/ocr.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('ocr')
-export class OcrController {}
