--- src/ocr/ocr.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { OcrService } from './ocr.service';
-
-describe('OcrService', () => {
-  let service: OcrService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [OcrService],
-    }).compile();
-
-    service = module.get<OcrService>(OcrService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
