+++ src/posts/posts.module.ts
@@ -5,7 +5,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { CommentsEntity } from '../comments/entities/comments.entity';
 import { CommentsModule } from '../comments/comments.module';
 import { PostsEntity } from './entities/base-posts.entity';
-import { ReportPostsEntity } from '../admin/entities/report-posts.entity';
+import { ReportPostsEntity } from '../reports/entities/report-posts.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([CommentsEntity, PostsEntity, ReportPostsEntity])],
