+++ src/posts/posts.service.ts
@@ -14,7 +14,7 @@ import { PostsEntity } from './entities/base-posts.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { IUserWithoutPassword } from '../auth/interfaces/session-decorator.interface';
-import { ReportPostsEntity } from '../admin/entities/report-posts.entity';
+import { ReportPostsEntity } from '../reports/entities/report-posts.entity';
 import { BasePostDto } from './dto/base-post.dto';
 import { ReportPostDto } from './dto/report-post.dto';
 import { EReportReason } from '../admin/enums';
