+++ src/users/entities/users.entity.ts
@@ -5,8 +5,8 @@ import { RepliesEntity } from '../../replies/entities/replies.entity';
 import { ScrapsEntity } from '../../scraps/entities/scraps.entity';
 import { Column, CreateDateColumn, Entity, OneToMany, PrimaryGeneratedColumn } from 'typeorm';
 import { LoginsEntity } from 'src/auth/entities/logins.entity';
-import { ReportPostsEntity } from '../../admin/entities/report-posts.entity';
-import { ReportCommentsEntity } from '../../admin/entities/report-comments.entity';
+import { ReportPostsEntity } from '../../reports/entities/report-posts.entity';
+import { ReportCommentsEntity } from '../../reports/entities/report-comments.entity';
 
 @Entity('users')
 export class UsersEntity {
